Select matches more carefully.
Only recognize matches that are at least "blocksize * 2" bytes.
Sometimes the "next pass" will find a better-suited match.

Sample data sets:
1: "*-12(Benefits for Organizations*-12(Benefits for Countries..."
2: "-12(Benefits for Organizations-12(Benefits for Countries..."

Without this patch, and with a blocksize of 16, the following blocks
are hashed:
"*-12(Benefits fo", "r Organizations*", "-12(Benefits for"
This causes the front of dataset 2 to be linked against the third
hashed block of dataset 1. (BAD!)

With this patch, the hashes match, but the match won't be recognized
because it's only 17 bytes (needs to be 32).
The next pass hashes the following blocks:
"*-12(Ben", "efits fo", "r Organi", "zations*", "-12(Bene", "fits for"
Now there are two reliable hashes to latch onto.