file.h: limit maximum amount of file I/O by 1MB
The problem was observed when I tried to run
bdelta.exe --all-in-ram on 32-bit windows on
network-mounted files.

fread(size=170MB) failed with 'out of memory' there.

I think it is a result of network-attached drives is
implemented in userspace or calling process which leads
to massive memory overhead when reading/writing large
chunks of data.

Fixed it by limiting I/O on 1MB size. It should I/O patterns
slightly better for fuse-mounted linux filesystems as well.

Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>